requirements.txt
selenium
webdriver-manager
pillow
requests
playwright
